Add export, some fixups to Wpf

+++ src/SerialLoops.Lib/Items/SfxItem.cs
@@ -5,10 +5,10 @@ namespace SerialLoops.Lib.Items
 {
     public class SfxItem : Item
     {
-        public int Index { get; set; }
+        public short Index { get; set; }
         public SfxEntry Entry { get; set; }
 
-        public SfxItem(SfxEntry entry, string name, int index) : base(name, ItemType.SFX)
+        public SfxItem(SfxEntry entry, string name, short index) : base(name, ItemType.SFX)
         {
             Entry = entry;
             Index = index;
